Remove old component files.
