--- src/components/CreditLogos-Composition.vue
@@ -1,44 +1,0 @@
-<template>
-  <div id="logo-credits" :style="cssVars">
-    <div id="icons-container">
-      <a href="https://www.cosmicds.cfa.harvard.edu/" target="_blank" rel="noopener noreferrer"
-        ><img alt="CosmicDS Logo" src="https://raw.githubusercontent.com/cosmicds/minids/main/assets/cosmicds_logo_for_dark_backgrounds.png"
-      /></a>
-      <a href="https://worldwidetelescope.org/home/" target="_blank" rel="noopener noreferrer"
-        ><img alt="WWT Logo" src="https://raw.githubusercontent.com/cosmicds/minids/main/assets/logo_wwt.png"
-      /></a>
-      <a href="https://science.nasa.gov/learners" target="_blank" rel="noopener noreferrer" class="pl-1"
-        ><img alt="SciAct Logo" src="https://raw.githubusercontent.com/cosmicds/minids/main/assets/logo_sciact.png"
-      /></a>
-      <a href="https://nasa.gov/" target="_blank" rel="noopener noreferrer" class="pl-1"
-        ><img alt="SciAct Logo" src="https://raw.githubusercontent.com/cosmicds/minids/main/assets/NASA_Partner_color_300_no_outline.png"
-      /></a>
-    </div>
-  </div>
-</template>
-
-<script setup lang="ts">
-import { computed } from "vue";
-
-export interface CreditLogosProps {
-  visible?: boolean;
-  logoSize?: string;
-}
-
-const props = withDefaults(defineProps<CreditLogosProps>(), {
-  visible: true,
-  logoSize: "5vmin"
-});
-
-const cssVars = computed(() => {
-  return {
-    "--logo-size": props.logoSize,
-  };
-});
-</script>
-
-<style scoped>
-#logo-credits img {
-  height: var(--logo-size);
-}
-</style>
