+++ src/components/CreditLogos.vue
@@ -17,37 +17,27 @@
   </div>
 </template>
 
-<script lang="ts">
-import { defineComponent } from "vue";
+<script setup lang="ts">
+import { computed } from "vue";
 
-export default defineComponent({
+export interface CreditLogosProps {
+  visible?: boolean;
+  logoSize?: string;
+}
 
-  props: {
-    visible: {
-      type: Boolean,
-      default: true
-    },
-    logoSize: {
-      type: String,
-      default: "5vmin"
-    }
-  },
+const props = withDefaults(defineProps<CreditLogosProps>(), {
+  visible: true,
+  logoSize: "5vmin"
+});
 
-  computed: { 
-    isMobile() {
-      return (/Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i.test(navigator.userAgent));
-    },
-    cssVars() {
-      return {
-        "--logo-size": this.logoSize,
-      };
-    }
-  },
-
+const cssVars = computed(() => {
+  return {
+    "--logo-size": props.logoSize,
+  };
 });
 </script>
 
-<style>
+<style scoped>
 #logo-credits img {
   height: var(--logo-size);
 }
