+++ src/components/IconButton.vue
@@ -40,102 +40,109 @@
   </v-tooltip>
 </template>
 
-<script lang="ts">
-import { defineComponent } from "vue";
-import { VTooltip } from "vuetify/components/VTooltip";
-import { VIcon } from "vuetify/components/VIcon";
+
+<script setup lang="ts">
+import { computed, ref, useAttrs } from "vue";
 import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome";
+import { VIcon } from "vuetify/components/VIcon";
+import { VTooltip } from "vuetify/components/VTooltip";
 
 // We need to do this because FontAwesome doesn't export the prop types
 type FontAwesomeIconProps = InstanceType<typeof FontAwesomeIcon>["$props"];
-
 type SizeType = Extract<FontAwesomeIconProps, 'size'>;
 
-export default defineComponent({
+interface IconButtonProps {
+  modelValue?: boolean;
+  faIcon?: string;
+  mdIcon?: string;
+  color?: string;
+  focusColor?: string;
+  backgroundColor?: string;
+  boxShadow?: boolean;
+  border?: boolean;
+  longPressTimeMs?: number;
+  tooltipText?: string;
+  tooltipLocation?: string;
+  tooltipOnClick?: boolean;
+  tooltipOnFocus?: boolean;
+  tooltipOnHover?: boolean;
+  tooltipOffset?: string | number;
+  showTooltip?: boolean;
+  faSize?: SizeType;
+  mdSize?: string;
+}
 
-  components: {
-    'v-tooltip': VTooltip,
-    'v-icon': VIcon,
-    'font-awesome-icon': FontAwesomeIcon
-  },
 
-  props: {
-    modelValue: { type: Boolean, required: false },
-    faIcon: { type: String, required: false, default: null},
-    mdIcon: { type: String, required: false, default: null},
-    color: { type: String, default: "#ffffff" },
-    focusColor: { type: String, default: "#ffffff" },
-    backgroundColor: { type: String, default: "#040404" },
-    boxShadow: { type: Boolean, default: true },
-    border: { type: Boolean, default: true },
-    longPressTimeMs: { type: Number, default: 500 },
-    tooltipText: { type: String, required: false },
-    tooltipLocation: { type: String, default: "start" },
-    tooltipOnClick: { type: Boolean, default: false },
-    tooltipOnFocus: { type: Boolean, default: false },
-    tooltipOnHover: { type: Boolean, default: true },
-    tooltipOffset: { type: [String, Number], default: 0 },
-    showTooltip: { type: Boolean, default: true },
-    faSize: { type: String as SizeType, default: 'lg' }, // fa: lg, md: 20px
-    mdSize: { type: String, default: '1.25em' } // fa: lg, md: 20px
-  },
-
-  methods: {
-    updateValue() {
-      if (this.modelValue === undefined) { return; }
-      this.$emit('update:modelValue', !this.modelValue); 
-    },
-
-    handleAction() {
-      this.updateValue();
-      this.$emit('activate');
-    },
-
-    handleTouchStart() {
-      this.longPressTimeout = setTimeout(() => {
-        this.tooltip = true;
-      }, this.longPressTimeMs);
-    },
-
-    handleTouchEnd() {
-      if (this.longPressTimeout) {
-        clearTimeout(this.longPressTimeout);
-        this.longPressTimeout = null;
-      }
-      this.tooltip = false;
-    }
-  },
-
-  data() {
-    return {
-      tooltip: false,
-      longPressTimeout: null as ReturnType<typeof setTimeout> | null
-    };
-  },
-
-  // Since our colors are used in compound values like e.g. box-shadows,
-  // we need to directly bind to CSS variables
-  computed: {
-    cssVars() {
-      return {
-        "--color": this.color,
-        "--background-color": this.backgroundColor,
-        "--focus-color": this.focusColor,
-        "--active-shadow": this.boxShadow ? this.color : 'transparent',
-        "--focus-shadow": this.boxShadow ? this.focusColor : 'transparent',
-      };
-    },
-
-    buttonID() {
-      const id = this.$attrs['id'];
-      return id ? `${id}-button` : null;
-    }
-  }
-
+const props = withDefaults(defineProps<IconButtonProps>(), {
+  color: "#ffffff",
+  focusColor: "#ffffff",
+  backgroundColor: "#040404",
+  boxShadow: true,
+  border: true,
+  longPressTimeMs: 500,
+  tooltipLocation: "start",
+  tooltipOnClick: false,
+  tooltipOnFocus: false,
+  tooltipOnHover: true,
+  tooltipOffset: 0,
+  showTooltip: true,
+  faSize: "lg",
+  mdSize: "1.25em"
 });
+
+const emit = defineEmits<{
+  "update:modelValue": [active: boolean],
+  "activate": [activate?: null]
+}>();
+
+
+type TimeoutType = ReturnType<typeof setTimeout>;
+const tooltip = ref(false);
+const longPressTimeout = ref<null | TimeoutType>(null);
+
+// Since our colors are used in compound values like e.g. box-shadows,
+// we need to directly bind to CSS variables
+const cssVars = computed(() => {
+  return {
+    "--color": props.color,
+    "--background-color": props.backgroundColor,
+    "--focus-color": props.focusColor,
+    "--active-shadow": props.boxShadow ? props.color : "transparent",
+    "--focus-shadow": props.boxShadow ? props.focusColor : "transparent",
+  };
+});
+
+const buttonID = computed(() => {
+  const attrs = useAttrs();
+  const id = attrs['id'];
+  return id ? `${id}-button`: null;
+});
+
+function updateValue() {
+  emit("update:modelValue", !props.modelValue);
+}
+
+function handleAction() {
+  updateValue();
+  emit('activate');
+}
+
+function handleTouchStart() {
+  longPressTimeout.value = setTimeout(() => {
+    tooltip.value = true;
+  }, props.longPressTimeMs);
+}
+
+function handleTouchEnd() {
+  if (longPressTimeout.value) {
+    clearTimeout(longPressTimeout.value);
+    longPressTimeout.value = null;
+  }
+  tooltip.value = false;
+}
 </script>
 
-<style lang="less">
+<style scoped lang="less">
 .icon-wrapper {
   color: var(--color);
   border-color: var(--color);
