--- src/components/MiniDSBase-Composition.ts
@@ -1,1 +1,0 @@
-
